Move bridge definitions
Move bridge definitions to their specific folder.
