+++ relays/bin-substrate/src/chains/rialto_headers_to_millau.rs
@@ -16,6 +16,7 @@
 
 //! Rialto-to-Millau headers sync entrypoint.
 
+use crate::cli::bridge::{CliBridgeBase, MessagesCliBridge, RelayToRelayHeadersCliBridge};
 use substrate_relay_helper::finality::{
 	engine::Grandpa as GrandpaFinalityEngine, DirectSubmitGrandpaFinalityProofCallBuilder,
 	SubstrateFinalitySyncPipeline,
@@ -36,4 +37,22 @@ impl SubstrateFinalitySyncPipeline for RialtoFinalityToMillau {
 		millau_runtime::RialtoGrandpaInstance,
 	>;
 	type TransactionSignScheme = relay_millau_client::Millau;
+}
+
+//// `Rialto` to `Millau` bridge definition.
+pub struct RialtoToMillauCliBridge {}
+
+impl CliBridgeBase for RialtoToMillauCliBridge {
+	type Source = relay_rialto_client::Rialto;
+	type Target = relay_millau_client::Millau;
+}
+
+impl RelayToRelayHeadersCliBridge for RialtoToMillauCliBridge {
+	type Finality = RialtoFinalityToMillau;
+}
+
+impl MessagesCliBridge for RialtoToMillauCliBridge {
+	const ESTIMATE_MESSAGE_FEE_METHOD: &'static str =
+		bp_millau::TO_MILLAU_ESTIMATE_MESSAGE_FEE_METHOD;
+	type MessagesLane = crate::chains::rialto_messages_to_millau::RialtoMessagesToMillau;
 }
