+++ relays/bin-substrate/src/chains/westend_headers_to_millau.rs
@@ -16,6 +16,7 @@
 
 //! Westend-to-Millau headers sync entrypoint.
 
+use crate::cli::bridge::{CliBridgeBase, RelayToRelayHeadersCliBridge};
 use substrate_relay_helper::finality::{
 	engine::Grandpa as GrandpaFinalityEngine, DirectSubmitGrandpaFinalityProofCallBuilder,
 	SubstrateFinalitySyncPipeline,
@@ -36,4 +37,16 @@ impl SubstrateFinalitySyncPipeline for WestendFinalityToMillau {
 		millau_runtime::WestendGrandpaInstance,
 	>;
 	type TransactionSignScheme = relay_millau_client::Millau;
+}
+
+//// `Westend` to `Millau` bridge definition.
+pub struct WestendToMillauCliBridge {}
+
+impl CliBridgeBase for WestendToMillauCliBridge {
+	type Source = relay_westend_client::Westend;
+	type Target = relay_millau_client::Millau;
+}
+
+impl RelayToRelayHeadersCliBridge for WestendToMillauCliBridge {
+	type Finality = WestendFinalityToMillau;
 }
