+++ relays/bin-substrate/src/cli/bridge.rs
@@ -111,106 +111,3 @@ pub trait MessagesCliBridge: CliBridgeBase {
 		RelayStrategy = MixStrategy,
 	>;
 }
-
-//// `Millau` to `Rialto` bridge definition.
-pub struct MillauToRialtoCliBridge {}
-
-impl CliBridgeBase for MillauToRialtoCliBridge {
-	type Source = relay_millau_client::Millau;
-	type Target = relay_rialto_client::Rialto;
-}
-
-impl RelayToRelayHeadersCliBridge for MillauToRialtoCliBridge {
-	type Finality = crate::chains::millau_headers_to_rialto::MillauFinalityToRialto;
-}
-
-impl MessagesCliBridge for MillauToRialtoCliBridge {
-	const ESTIMATE_MESSAGE_FEE_METHOD: &'static str =
-		bp_rialto::TO_RIALTO_ESTIMATE_MESSAGE_FEE_METHOD;
-	type MessagesLane = crate::chains::millau_messages_to_rialto::MillauMessagesToRialto;
-}
-
-//// `Rialto` to `Millau` bridge definition.
-pub struct RialtoToMillauCliBridge {}
-
-impl CliBridgeBase for RialtoToMillauCliBridge {
-	type Source = relay_rialto_client::Rialto;
-	type Target = relay_millau_client::Millau;
-}
-
-impl RelayToRelayHeadersCliBridge for RialtoToMillauCliBridge {
-	type Finality = crate::chains::rialto_headers_to_millau::RialtoFinalityToMillau;
-}
-
-impl MessagesCliBridge for RialtoToMillauCliBridge {
-	const ESTIMATE_MESSAGE_FEE_METHOD: &'static str =
-		bp_millau::TO_MILLAU_ESTIMATE_MESSAGE_FEE_METHOD;
-	type MessagesLane = crate::chains::rialto_messages_to_millau::RialtoMessagesToMillau;
-}
-
-//// `Westend` to `Millau` bridge definition.
-pub struct WestendToMillauCliBridge {}
-
-impl CliBridgeBase for WestendToMillauCliBridge {
-	type Source = relay_westend_client::Westend;
-	type Target = relay_millau_client::Millau;
-}
-
-impl RelayToRelayHeadersCliBridge for WestendToMillauCliBridge {
-	type Finality = crate::chains::westend_headers_to_millau::WestendFinalityToMillau;
-}
-
-//// `Millau` to `RialtoParachain`  bridge definition.
-pub struct MillauToRialtoParachainCliBridge {}
-
-impl CliBridgeBase for MillauToRialtoParachainCliBridge {
-	type Source = relay_millau_client::Millau;
-	type Target = relay_rialto_parachain_client::RialtoParachain;
-}
-
-impl RelayToRelayHeadersCliBridge for MillauToRialtoParachainCliBridge {
-	type Finality =
-		crate::chains::millau_headers_to_rialto_parachain::MillauFinalityToRialtoParachain;
-}
-
-impl MessagesCliBridge for MillauToRialtoParachainCliBridge {
-	const ESTIMATE_MESSAGE_FEE_METHOD: &'static str =
-		bp_rialto_parachain::TO_RIALTO_PARACHAIN_ESTIMATE_MESSAGE_FEE_METHOD;
-	type MessagesLane =
-		crate::chains::millau_messages_to_rialto_parachain::MillauMessagesToRialtoParachain;
-}
-
-//// `RialtoParachain` to `Millau` bridge definition.
-pub struct RialtoParachainToMillauCliBridge {}
-
-impl CliBridgeBase for RialtoParachainToMillauCliBridge {
-	type Source = relay_rialto_parachain_client::RialtoParachain;
-	type Target = relay_millau_client::Millau;
-}
-
-impl ParachainToRelayHeadersCliBridge for RialtoParachainToMillauCliBridge {
-	type SourceRelay = relay_rialto_client::Rialto;
-	type ParachainFinality = crate::chains::rialto_parachains_to_millau::RialtoParachainsToMillau;
-	type RelayFinality = crate::chains::rialto_headers_to_millau::RialtoFinalityToMillau;
-}
-
-impl MessagesCliBridge for RialtoParachainToMillauCliBridge {
-	const ESTIMATE_MESSAGE_FEE_METHOD: &'static str =
-		bp_millau::TO_MILLAU_ESTIMATE_MESSAGE_FEE_METHOD;
-	type MessagesLane =
-		crate::chains::rialto_parachain_messages_to_millau::RialtoParachainMessagesToMillau;
-}
-
-//// `WestendParachain` to `Millau` bridge definition.
-pub struct WestmintToMillauCliBridge {}
-
-impl ParachainToRelayHeadersCliBridge for WestmintToMillauCliBridge {
-	type SourceRelay = relay_westend_client::Westend;
-	type ParachainFinality = crate::chains::westend_parachains_to_millau::WestendParachainsToMillau;
-	type RelayFinality = crate::chains::westend_headers_to_millau::WestendFinalityToMillau;
-}
-
-impl CliBridgeBase for WestmintToMillauCliBridge {
-	type Source = relay_westend_client::Westmint;
-	type Target = relay_millau_client::Millau;
-}
